% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Account-class}
\alias{Account-class}
\alias{Account}
\title{An S4 class to represent webhose results.}
\description{
An S4 class to represent webhose results.
}
\section{Slots}{

\describe{
\item{\code{balance}}{A length-one numeric vector}
}}

