% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wh_token.R
\name{token}
\alias{token}
\alias{wh_token}
\alias{wh_get_token}
\title{Setup}
\usage{
wh_token(token)

wh_get_token(quiet = !interactive())
}
\arguments{
\item{token}{your token, see details.}

\item{quiet}{if \code{FALSE} does not return useful information to console.}
}
\description{
Specify token to be used in all subsequent calls.
}
\details{
\enumerate{
  \item{Create a \href{https://webhose.io}{webhose.io} account.}
  \item{Visit your \href{https://webhose.io/dashboard}{dashboard}.}
  \item{Use the "Active API Key" at the bottom as \code{token} parameter.}
}
}
\note{
You can specify the \code{WEBHOSER_TOKEN} environment variable, likely in your code{.Renviron}.
}
\examples{
wh_token("xXX-x0X0xX0X-00X")

}
