% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{wh_collect}
\alias{wh_collect}
\alias{wh_collect.webhoser}
\title{Collect results}
\usage{
wh_collect(wh, flatten = FALSE)

\method{wh_collect}{webhoser}(wh, flatten = FALSE)
}
\arguments{
\item{wh}{object of class \code{webhoser} as returned by \code{\link{wh_news}}.}

\item{flatten}{whether to flatten the collected results.}
}
\value{
nested \code{data.frame} if \code{flatten = FALSE}. See
\href{https://docs.webhose.io/v1.0/docs/output-reference}{official documentation}
}
\description{
Collect results from \code{webhoser} object.
}
\details{
\code{flatten} ignores sentiment of entities.
}
\examples{
\dontrun{
token <- <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news(q = "R programming language") \%>\%
  wh_paginate(p = 1) \%>\%
  wh_collect -> rstats
}

}
