% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{wh_date}
\alias{wh_date}
\title{format date}
\usage{
wh_date(date)
}
\arguments{
\item{date}{date column, see examples.}
}
\value{
vector of class \code{POSIXct}.
}
\description{
Format dates returned by API, see details.
}
\details{
Webhose's servers are in Israel, this is used as timezone.
}
\examples{
\dontrun{
token <- wh_token("xXX-x0X0xX0X-00X")

wef <- wh_news(token, q = '"World Economic Forum"') \%>\%  # use quote marks!
  wh_collect() \%>\%
  dplyr::mutate(published = wh_date(published))# collect results
}

}
