% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{wh_reviews}
\alias{wh_reviews}
\title{Get Reviews}
\usage{
wh_reviews(q, ts = (Sys.time() - (3 * 24 * 60 * 60)), sort = NULL,
  order = NULL, accuracy = NULL, highlight = NULL, latest = NULL,
  quiet = !interactive())
}
\arguments{
\item{q}{a string query containing the filters that define which posts will be returned.}

\item{ts}{The "ts" (timestamp) parameter is telling the system to return results that were
crawled after this timestamp (\code{POSIXct} or \code{POSIXlt}).}

\item{sort}{by default (when the sort parameter isn't specified)
the results are sorted by the recommended order of crawl date.
See details for valid values.}

\item{order}{If you choose to order the posts by any of the numeric \code{sort} values,
you can choose in what order you want to get them: \code{asc} (default) or \code{desc}.}

\item{accuracy}{return only posts with high extraction accuracy, but removes about 30% of
the total matching posts (with lower confidence).}

\item{highlight}{return the fragments in the post that matched the textual Boolean query.
The matched keywords will be surrounded by \code{<em/>} tags.}

\item{latest}{this will return the latest 100 crawled posts matching your query (**NOT** recommended).}

\item{quiet}{if \code{FALSE} does not return useful information to console.}
}
\value{
object of class \code{webhoser}
}
\description{
Get access to structured online reviews.
}
\details{
Valid sort values
\itemize{
  \item{relevancy}
  \item{social.facebook.likes}
  \item{social.facebook.shares}
  \item{social.facebook.comments}
  \item{social.gplus.shares}
  \item{social.pinterest.shares}
  \item{social.linkedin.shares}
  \item{social.stumbledupon.shares}
  \item{social.vk.shares}
  \item{replies_count}
  \item{participants_count}
  \item{spam_score}
  \item{performance_score}
  \item{published}
  \item{thread.published}
  \item{domain_rank}
  \item{ord_in_thread}
  \item{rating}
}

See \href{https://docs.webhose.io/v1.0/docs/filters-reference}{official documentation} for valid filters.
}
\examples{
\dontrun{
token <- wh_token("xXX-x0X0xX0X-00X")

rstats <- wh_reviews(q = '"ipad OR ipod"') \%>\%  # use quote marks!
  wh_collect() # collect results
}

}
