% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wh_token.R
\name{wh_token}
\alias{wh_token}
\title{Setup webhose token}
\usage{
wh_token(token)
}
\arguments{
\item{token}{your token, see details.}
}
\description{
Setup webhose token
}
\details{
\enumerate{
  \item{Create a \href{https://webhose.io}{webhose.io} account.}
  \item{Visit your \href{https://webhose.io/dashboard}{dashboard}.}
  \item{Use the "Active API Key" at the bottom as \code{token} parameter.}
}
}
\examples{
token <- wh_token("xXX-x0X0xX0X-00X")

}
