% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcasts.R
\name{wh_broadcasts}
\alias{wh_broadcasts}
\title{Get broadcasts}
\usage{
wh_broadcasts(token, q, ts = (Sys.time() - (3 * 24 * 60 * 60)),
  highlight = NULL, latest = NULL, quiet = !interactive())
}
\arguments{
\item{token}{your token as returned by \code{\link{wh_token}}.}

\item{q}{a string query containing the filters that define which transcript lines will be returned.}

\item{ts}{The "ts" (timestamp) parameter is telling the system to return results that were
crawled after this timestamp (\code{POSIXct} or \code{POSIXlt}).}

\item{highlight}{return the fragments in the review that matched the textual Boolean query.
The matched keywords will be surrounded by \code{<em/>} tags.}

\item{latest}{this will return the latest 100 crawled posts matching your query (**NOT** recommended).}

\item{quiet}{if \code{FALSE} does not return useful information to console.}
}
\description{
Get access to structured trasncripts data from broadcasts.
}
\details{
See \href{https://docs.webhose.io/docs/filters-reference-4}{official documentation} for valid filters.
}
\examples{
\dontrun{
token <- wh_token("xXX-x0X0xX0X-00X")

wef <- wh_broadcasts(token, q = '"World Economic Forum"') \%>\%  # use quote marks!
  wh_collect()

token \%>\%
  wh_broadcasts(token, q = 'Davos OR "World Economic Forum" OR WEF') \%>\%
  wh_paginate(p = 2) # 2 additional pages of results
  wh_collect() -> davos
}

}
