% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{wh_paginate}
\alias{wh_paginate}
\alias{wh_paginate.webhoser}
\title{Paginate}
\usage{
wh_paginate(wh, p = Inf, quiet = FALSE)

\method{wh_paginate}{webhoser}(wh, p = Inf, quiet = FALSE)
}
\arguments{
\item{wh}{object of class \code{webhoser} as returned by \code{\link{wh_news_filter}}.}

\item{p}{number of pages defaults to \code{Inf}.}

\item{quiet}{if \code{FALSE} does not return useful information to console.}
}
\description{
Paginate through query.
}
\examples{
\dontrun{
token <- <- wh_token("xXX-x0X0xX0X-00X")

token \%>\%
  wh_news_filter(q = "Barack Obama") \%>\%
  wh_paginate(p = 3) -> barack
}

}
